Move exceptions into sprockets_amqp.exceptions.
