+++ docs/api.rst
@@ -1,4 +1,9 @@
 API
----
+===
 .. automodule:: sprockets.mixins.amqp
+   :members:
+
+Exceptions
+----------
+.. automodule:: sprockets_amqp.exceptions
    :members:
