+++ sprockets_amqp/__init__.py
@@ -1,0 +1,2 @@
+version_info = (2, 2, 0)
+version = '.'.join(str(v) for v in version_info)
